+++ src/components/menu/index.tsx
@@ -16,20 +16,33 @@ import ModalPadraoDeSacola from "../modalPadraoDeSacola";
 import ModalEnvioErros from "../modalEnvioErros";
 import { useData } from "@/providers/dataProvider";
 import ModalEnvioPlanilha from "../modalEnvioPlanilha";
+import { toast } from "react-hot-toast";
+import iconRobo from "../../../public/Icon_Robo.svg";
 
 const Menu = () => {
   const { setContent, showModal, reverseModal } = useModal();
-  const { setExcelFile } =useData()
+  const { setExcelFile, errorsLog } = useData();
 
-  const handleFileUpload = (event :  React.ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0]
-    if (file) {
-      setExcelFile(file)
-      setContent(<ModalEnvioPlanilha/>);
-      showModal();
+  const handleFileUpload = (event: React.ChangeEvent<HTMLInputElement>) => {
+    if (!errorsLog) {
+      const file = event.target.files?.[0];
+      if (file) {
+        setExcelFile(file);
+        setContent(<ModalEnvioPlanilha />);
+        showModal();
+      }
     }
-
-  }
+     toast("REVISE OS ERROS ANTES DE ENVIAR UMA NOVA PLANILHA", {
+       icon: <Image src={iconRobo} alt="Supp" className="h-[3rem] w-[3rem]" />,
+       style: {
+         borderRadius: "50px",
+         background: "#F4F3F7",
+         color: "#3C2F58",
+         fontSize: "1.3rem",
+         fontWeight: "bolder",
+       },
+     });
+  };
 
   return (
     <div className="flex flex-col justify-between fixed w-[21.2rem] h-screen drop-shadow-xl bg-branco-primario text-roxo-primario text-[1.8rem] animate-showModalAnimation z-5">
